change to ssa

+++ g2o/apps/g2o_viewer/g2o_qglviewer.cpp
@@ -43,14 +43,14 @@ namespace {
     public:
       StandardCamera() : _standard(true) {};
 
-      float zNear() const {
+      qreal zNear() const {
         if (_standard) 
           return 0.001f; 
         else 
           return Camera::zNear(); 
       }
 
-      float zFar() const
+      qreal zFar() const
       {  
         if (_standard) 
           return 10000.0f;
